Windows port: experimental conv3d: fix narrowing in Shape init
- prepare_conv3d_weights.cpp: static_cast<uint32_t>(-1) for the
  reshape's infer-dim placeholder passed into ttnn::Shape (C2398
  narrowing of int -1 to uint32_t), matching the (uint32_t)-1
  sentinel idiom used elsewhere in the tree.

+++ ttnn/cpp/ttnn/operations/experimental/conv3d/prepare_conv3d_weights.cpp
@@ -193,7 +193,7 @@ Tensor prepare_conv3d_weights(
         ttnn::reshape(prepare_weights, ttnn::Shape{kD, kH, kW, num_C_in_blocks, C_in_block, out_channels});
     ttsl::SmallVector<int64_t> dims_2 = {3, 0, 1, 2, 4, 5};
     prepare_weights = ttnn::permute(prepare_weights, dims_2);
-    prepare_weights = ttnn::reshape(prepare_weights, ttnn::Shape{-1, out_channels});
+    prepare_weights = ttnn::reshape(prepare_weights, ttnn::Shape{static_cast<uint32_t>(-1), out_channels});
     return prepare_weights;
 }
